1. Finger print for Pay and Approve bills 2. Fixed Line Item display issue in iOS 7/8 3. Fixed attach pic to existing bill/invoice error

+++ BDC/BOSelectorViewController.m
@@ -125,7 +125,7 @@
 }
 
 - (void)performNewSegueForCell:(UIButton *)button {
-    int i = button.tag / 10;
+    NSInteger i = button.tag / 10;
     int j = button.tag % 10;
     NSString *segueId = AttachToNewSegues[i][j];
     [self performSegueWithIdentifier:segueId sender:self];
