+++ BDC/EditItemViewController.m
@@ -328,7 +328,7 @@ enum ItemInfoType {
     return cell;
 }
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath {
+- (double)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath {
     if (indexPath.section == 0) {
         return CELL_HEIGHT;
     } else {
